Time to start over with latest FB SDK!
Replacing HEAD with the latest version of the FB SDK, overwriting our previous patch. If it turns out we need that, we can reapply it here, or make a branch for older projects.

For the record, this is effectively a 'git merge -s theirs'; I did it by following these instructions:
http://www.seanius.net/blog/2011/02/git-merge-s-theirs/